Configuration related code moved to separate namespace for clarity
