+++ ReasonCodeExample.XPathTools/ReasonCodeExample.XPathTools.csproj
@@ -168,7 +168,7 @@
     <Compile Include="VisualStudioIntegration\CaretPositionLineInfo.cs" />
     <Compile Include="SimpleXmlNamespaceResolver.cs" />
     <Compile Include="VisualStudioIntegration\XPathFormat.cs" />
-    <Compile Include="VisualStudioIntegration\XPathToolsDialogPage.cs">
+    <Compile Include="Configuration\XPathToolsDialogPage.cs">
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="VisualStudioIntegration\XPathToolsPackage.cs" />
@@ -184,8 +184,8 @@
     <Compile Include="Workbench\XPathWorkbenchWindow.cs" />
     <Compile Include="Writers\TrimCommandTextFormatter.cs" />
     <Compile Include="Writers\CommandTextFormatter.cs" />
-    <Compile Include="VisualStudioIntegration\SerializableConverter.cs" />
-    <Compile Include="IConfiguration.cs" />
+    <Compile Include="Configuration\SerializableConverter.cs" />
+    <Compile Include="Configuration\IConfiguration.cs" />
     <Compile Include="Workbench\SearchResult.cs" />
     <Compile Include="Workbench\SearchResultFactory.cs" />
     <Compile Include="Workbench\XPathWorkbench.cs">
